[draft] complex & refactor
